feat: add ui to docs
