+++ packages/docs/.storybook/main.ts
@@ -1,0 +1,37 @@
+import { dirname, join } from "path";
+import { StorybookConfig } from "storybook-framework-qwik";
+
+const config: StorybookConfig = {
+  addons: [
+    getAbsolutePath("@storybook/addon-links"),
+    getAbsolutePath("@storybook/addon-essentials"),
+  ],
+
+  framework: {
+    name: getAbsolutePath("storybook-framework-qwik"),
+  },
+
+  core: {
+    renderer: getAbsolutePath("storybook-framework-qwik"),
+  },
+
+  stories: [
+    // ...rootMain.stories,
+    "../src/components/**/*.stories.mdx",
+    "../src/components/**/*.stories.@(js|jsx|ts|tsx)",
+  ],
+
+  viteFinal: async (config: any) => {
+    return config;
+  },
+
+  docs: {
+    autodocs: true,
+  },
+};
+
+export default config;
+
+function getAbsolutePath(value: string): any {
+  return dirname(require.resolve(join(value, "package.json")));
+}
