+++ packages/docs/package.json
@@ -11,6 +11,7 @@
   "private": true,
   "scripts": {
     "build": "qwik build",
+    "build-storybook": "storybook build",
     "build.client": "vite build",
     "build.preview": "vite build --ssr src/entry.preview.tsx",
     "build.server": "vite build -c adapters/static/vite.config.ts",
@@ -23,27 +24,42 @@
     "lint": "eslint \"src/**/*.ts*\"",
     "preview": "qwik build preview && vite preview --open",
     "start": "vite --open --mode ssr",
-    "test": "playwright test",
+    "storybook": "storybook dev -p 6006",
+    "test": "pnpm test.unit && pnpm test.e2e",
+    "test.e2e": "playwright test",
+    "test.unit": "vitest run components",
+    "test.unit.ui": "vitest --ui components",
     "qwik": "qwik"
   },
   "devDependencies": {
     "@builder.io/partytown": "^0.8.0",
-    "@builder.io/qwik": "^1.2.6",
-    "@builder.io/qwik-city": "^1.2.6",
-    "@playwright/test": "^1.35.1",
+    "@builder.io/qwik": "^1.2.7",
+    "@builder.io/qwik-city": "^1.2.7",
+    "@playwright/test": "^1.37.0",
+    "@qwikest/icons": "^0.0.8",
+    "@storybook/addon-essentials": "7.3.1",
+    "@storybook/addon-links": "7.3.1",
+    "@storybook/blocks": "7.3.1",
+    "@storybook/builder-vite": "7.3.1",
+    "@storybook/html": "7.3.1",
+    "@storybook/html-vite": "7.3.1",
     "@types/eslint": "8.40.2",
-    "@types/node": "^20.4.0",
+    "@types/node": "^20.5.0",
     "@typescript-eslint/eslint-plugin": "5.61.0",
     "@typescript-eslint/parser": "5.61.0",
+    "@vitest/ui": "^0.33.0",
     "eslint": "8.44.0",
-    "eslint-plugin-qwik": "^1.2.6",
+    "eslint-plugin-qwik": "^1.2.7",
+    "eslint-plugin-storybook": "^0.6.13",
     "prettier": "2.8.8",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "storybook": "7.3.1",
+    "storybook-framework-qwik": "0.2.3",
     "typescript": "5.1.6",
     "undici": "5.22.1",
-    "vite": "4.4.0",
-    "vite-tsconfig-paths": "4.2.0"
-  },
-  "dependencies": {
-    "@sb/ui": "workspace:^"
+    "vite": "4.4.9",
+    "vite-tsconfig-paths": "4.2.0",
+    "vitest": "^0.33.0"
   }
 }
