+++ packages/docs/src/components/example/example.tsx
