+++ packages/docs/src/routes/blog/index.tsx
@@ -1,6 +1,6 @@
 import { component$ } from "@builder.io/qwik";
-import { PostSummaryList } from "@sb/ui";
 import data from "./index.json";
+import { PostSummaryList } from "~/components/post-summary-list/post-summary-list";
 
 export default component$(() => {
   return (
