+++ packages/docs/src/routes/index.md
@@ -13,43 +13,43 @@ callToActionItems:
 
 <img src="https://google-account-photo.vercel.app/api/?account_id=101722346324226588907" style="width: 100px"/>
 
-大学時代に初めてWebAppを開発し, 目に見える形までアウトプットできたことに感動し, 今の職業を目指すようになる. ふとWebAppのアイデアが浮かんだときは, 素早く,安く, そしてお手軽に リリースできることを目指して個人開発をしている。
+大学時代に初めて WebApp を開発し, 目に見える形までアウトプットできたことに感動し, 今の職業を目指すようになる. ふと WebApp のアイデアが浮かんだときは, 素早く,安く, そしてお手軽に リリースできることを目指して個人開発をしている。
 
 ## 興味
 
-* Micro Frontends
-* Browser Engine
-* Rust・WebAssembly
-* Test Engineering
-* Workflow Engine
-* Google Cloud Platform
+- Micro Frontends
+- Browser Engine
+- Rust・WebAssembly
+- Test Engineering
+- Workflow Engine
+- Google Cloud Platform
 
 ## 経歴
 
-|日付|概要|
-|--|--|
-|2016/03|同志社大学 インテリジェント情報工学科 学士課程 卒業|
-|2016/04|株式会社テイジイエル 入社|
-|2018/07|株式会社テイジイエル 退社|
-|2018/08|株式会社MonotaRO 入社|
-|2021/12|株式会社MonotaRO 退社|
-|2022/01|株式会社マネーフォワード 入社|
-|2023/06|株式会社マネーフォワード 退社|
+| 日付    | 概要                                                |
+| ------- | --------------------------------------------------- |
+| 2016/03 | 同志社大学 インテリジェント情報工学科 学士課程 卒業 |
+| 2016/04 | 株式会社テイジイエル 入社                           |
+| 2018/07 | 株式会社テイジイエル 退社                           |
+| 2018/08 | 株式会社 MonotaRO 入社                              |
+| 2021/12 | 株式会社 MonotaRO 退社                              |
+| 2022/01 | 株式会社マネーフォワード 入社                       |
+| 2023/06 | 株式会社マネーフォワード 退社                       |
 
 ## 資格
 
-|日付|概要|
-|--|--|
-|2015/04|応用情報技術者試験 合格|
-|2017/09|Java SE 8 合格|
-|2023/05|TOEIC 640点|
+| 日付    | 概要                    |
+| ------- | ----------------------- |
+| 2015/04 | 応用情報技術者試験 合格 |
+| 2017/09 | Java SE 8 合格          |
+| 2023/05 | TOEIC 640 点            |
 
 ## 実績
 
-|日付|概要|詳細|備考|
-|--|--|--|--|
-|2019/02|GCP hands-on 優勝|[Developers Summit 2019](https://event.shoeisha.jp/devsumi/20190214/session/2015/)|[See certificate](https://res.cloudinary.com/silverbirder/image/upload/v1551278903/Accomplish%C2%ADments/developers_summit_2019_gcp_handson.jpg)|
-|2019/06|初登壇|[Developers Boost KANSAI ～関西U30エンジニアの登竜門～](https://event.shoeisha.jp/devboost/20190615/timetable#tt1810)|[Slide](https://www.slideshare.net/monotaro-itd-pr/ss-150331504)|
-|2019/09|初出版|[技術書典7『はじめてのWeb Components入門 -4つの基本機能から関連ライブラリまで-』](https://techbookfest.org/event/tbf07/circle/5117648689954816)|[Sold here](https://www.amazon.co.jp/dp/B08CY2QCFV/)|
-|2019/11|第三回 銭けっと サークル参加|[銭けっと](https://zeniket.jimdofree.com/)||
-|2019/12|初インタビュー|[Engineer Hub](https://employment.en-japan.com/engineerhub/entry/2019/12/19/103000)||
+| 日付    | 概要                         | 詳細                                                                                                                                                | 備考                                                                                                                                             |
+| ------- | ---------------------------- | --------------------------------------------------------------------------------------------------------------------------------------------------- | ------------------------------------------------------------------------------------------------------------------------------------------------ |
+| 2019/02 | GCP hands-on 優勝            | [Developers Summit 2019](https://event.shoeisha.jp/devsumi/20190214/session/2015/)                                                                  | [See certificate](https://res.cloudinary.com/silverbirder/image/upload/v1551278903/Accomplish%C2%ADments/developers_summit_2019_gcp_handson.jpg) |
+| 2019/06 | 初登壇                       | [Developers Boost KANSAI ～関西 U30 エンジニアの登竜門～](https://event.shoeisha.jp/devboost/20190615/timetable#tt1810)                             | [Slide](https://www.slideshare.net/monotaro-itd-pr/ss-150331504)                                                                                 |
+| 2019/09 | 初出版                       | [技術書典 7『はじめての Web Components 入門 -4 つの基本機能から関連ライブラリまで-』](https://techbookfest.org/event/tbf07/circle/5117648689954816) | [Sold here](https://www.amazon.co.jp/dp/B08CY2QCFV/)                                                                                             |
+| 2019/11 | 第三回 銭けっと サークル参加 | [銭けっと](https://zeniket.jimdofree.com/)                                                                                                          |                                                                                                                                                  |
+| 2019/12 | 初インタビュー               | [Engineer Hub](https://employment.en-japan.com/engineerhub/entry/2019/12/19/103000)                                                                 |                                                                                                                                                  |
