+++ packages/docs/src/routes/misc/artifact/index.md
@@ -2,49 +2,49 @@
 
 ## Web Site / App
 
-* [マイクロフロントエンド入門](https://intro-to-micro-frontends.netlify.app/)
-* [GAP(Gather the presentation)](https://silverbirder-gap.netlify.app/)
+- [マイクロフロントエンド入門](https://intro-to-micro-frontends.netlify.app/)
+- [GAP(Gather the presentation)](https://silverbirder-gap.netlify.app/)
 
 ## Github My App
 
-* [https://github.com/silverbirder/CaAT](https://github.com/silverbirder/CaAT)
-* [https://github.com/silverbirder/Cotlin](https://github.com/silverbirder/Cotlin)
-* [https://github.com/silverbirder/google-word-cloud](https://github.com/silverbirder/google-word-cloud)
-* [https://github.com/silverbirder/notify-message-by-google-home](https://github.com/silverbirder/notify-message-by-google-home)
-* [https://github.com/silverbirder/puppeteer-account-manager](https://github.com/silverbirder/puppeteer-account-manager)
-* [https://github.com/silverbirder/rMinc](https://github.com/silverbirder/rMinc)
-* [https://github.com/silverbirder/run_character](https://github.com/silverbirder/run_character)
-* [https://github.com/silverbirder/search-local-web](https://github.com/silverbirder/search-local-web)
-* [https://github.com/silverbirder/searchWordCloud](https://github.com/silverbirder/searchWordCloud)
-  * Chrome Web Store
-    * [searchWordCloud](https://chrome.google.com/webstore/detail/searchwordcloud/dbpmolojnmdfgggfnhlioepakmpjafal?hl=ja)
-* [https://github.com/silverbirder/silver-enigma](https://github.com/silverbirder/silver-enigma)
-* [https://github.com/silverbirder/tiqav2](https://github.com/silverbirder/tiqav2)
-* [https://github.com/silverbirder/zoom-meeting-creator](https://github.com/silverbirder/zoom-meeting-creator)
-* [https://github.com/silverbirder/Google-Account-Photo-API](https://github.com/silverbirder/Google-Account-Photo-API)
-* [https://github.com/silverbirder/chrome-extensions-tiktok-scraping-downloader](https://github.com/silverbirder/chrome-extensions-tiktok-scraping-downloader)
-* [https://github.com/silverbirder/ai-ghost-writer](https://github.com/silverbirder/ai-ghost-writer)
-  * Chrome Web Store
-    * [AI Ghostwriter](https://chrome.google.com/webstore/detail/ai-ghostwriter/hpcokeldeijnfmbbbjkedhnedjjbjmoa)
+- [https://github.com/silverbirder/CaAT](https://github.com/silverbirder/CaAT)
+- [https://github.com/silverbirder/Cotlin](https://github.com/silverbirder/Cotlin)
+- [https://github.com/silverbirder/google-word-cloud](https://github.com/silverbirder/google-word-cloud)
+- [https://github.com/silverbirder/notify-message-by-google-home](https://github.com/silverbirder/notify-message-by-google-home)
+- [https://github.com/silverbirder/puppeteer-account-manager](https://github.com/silverbirder/puppeteer-account-manager)
+- [https://github.com/silverbirder/rMinc](https://github.com/silverbirder/rMinc)
+- [https://github.com/silverbirder/run_character](https://github.com/silverbirder/run_character)
+- [https://github.com/silverbirder/search-local-web](https://github.com/silverbirder/search-local-web)
+- [https://github.com/silverbirder/searchWordCloud](https://github.com/silverbirder/searchWordCloud)
+  - Chrome Web Store
+    - [searchWordCloud](https://chrome.google.com/webstore/detail/searchwordcloud/dbpmolojnmdfgggfnhlioepakmpjafal?hl=ja)
+- [https://github.com/silverbirder/silver-enigma](https://github.com/silverbirder/silver-enigma)
+- [https://github.com/silverbirder/tiqav2](https://github.com/silverbirder/tiqav2)
+- [https://github.com/silverbirder/zoom-meeting-creator](https://github.com/silverbirder/zoom-meeting-creator)
+- [https://github.com/silverbirder/Google-Account-Photo-API](https://github.com/silverbirder/Google-Account-Photo-API)
+- [https://github.com/silverbirder/chrome-extensions-tiktok-scraping-downloader](https://github.com/silverbirder/chrome-extensions-tiktok-scraping-downloader)
+- [https://github.com/silverbirder/ai-ghost-writer](https://github.com/silverbirder/ai-ghost-writer)
+  - Chrome Web Store
+    - [AI Ghostwriter](https://chrome.google.com/webstore/detail/ai-ghostwriter/hpcokeldeijnfmbbbjkedhnedjjbjmoa)
 
 ## WebComponents
 
-* [@silverbirder/o-embed](https://www.webcomponents.org/element/Silver-birder/o-embed)
-* [@silverbirder/ogp-me](https://www.webcomponents.org/element/Silver-birder/ogp-me)
+- [@silverbirder/o-embed](https://www.webcomponents.org/element/Silver-birder/o-embed)
+- [@silverbirder/ogp-me](https://www.webcomponents.org/element/Silver-birder/ogp-me)
 
 ## SpreadSheet
 
-* [POV](https://docs.google.com/spreadsheets/d/e/2PACX-1vSpbSeaOPVSKyi36bwbBXQ56DbXNzLEp-anI4PHfXps4pa7gWUMDGHjNmVy1gl945o4aNGCszPWxcKm/pubhtml)
+- [POV](https://docs.google.com/spreadsheets/d/e/2PACX-1vSpbSeaOPVSKyi36bwbBXQ56DbXNzLEp-anI4PHfXps4pa7gWUMDGHjNmVy1gl945o4aNGCszPWxcKm/pubhtml)
 
 ## Subscribe
 
-* [DeepL Pro](https://www.deepl.com/translator)
-* [ChatGPT Plus](https://openai.com/blog/chatgpt-plus)
-* [Google One 2TB](https://one.google.com/about/plans)
-* [Amazon Prime](https://www.amazon.co.jp/amazonprime)
-* [dアニメストア](https://animestore.docomo.ne.jp/animestore)
-* ~~[Twitter Blue](https://help.twitter.com/ja/using-twitter/twitter-blue)~~
+- [DeepL Pro](https://www.deepl.com/translator)
+- [ChatGPT Plus](https://openai.com/blog/chatgpt-plus)
+- [Google One 2TB](https://one.google.com/about/plans)
+- [Amazon Prime](https://www.amazon.co.jp/amazonprime)
+- [d アニメストア](https://animestore.docomo.ne.jp/animestore)
+- ~~[Twitter Blue](https://help.twitter.com/ja/using-twitter/twitter-blue)~~
 
 ## Slide
 
-* [至宝のミスタードーナツ](https://gamma.app/docs/--gsja1jdqk543srm?token=jg0q3n5ipafqw5v)
+- [至宝のミスタードーナツ](https://gamma.app/docs/--gsja1jdqk543srm?token=jg0q3n5ipafqw5v)
