+++ packages/docs/src/routes/misc/watch/index.md
@@ -1,58 +1,60 @@
 # Misc >> Watch
+
 ## RSS
+
 ### Tech Company Blog
 
- * [Chromium Blog](http://blog.chromium.org/atom.xml)
- * [Google Developers Blog](http://feeds.feedburner.com/GDBcode)
- * [AWS News Blog](https://aws.amazon.com/jp/blogs/aws/feed/)
- * [Amazon Web Services Blog](https://aws.amazon.com/jp/blogs/news/feed/)
- * [Cloudflare Blog](https://blog.cloudflare.com/rss/)
- * [Atlassian Developer Blog](https://blog.developer.atlassian.com/feed/)
- * [Google Chrome Blog](https://blog.google/products/chrome/rss)
- * [Heroku](https://blog.heroku.com/engineering/feed)
- * [Twitter Blog](https://blog.twitter.com/engineering/en_us/blog.rss)
- * [WHATWG Blog](https://blog.whatwg.org/feed)
- * [Google Cloud Platform Blog Japan](https://cloudblog.withgoogle.com/ja/products/gcp/rss/)
- * [Google Developers Japan](https://developers-jp.googleblog.com/atom.xml)
- * [Uber Engineering Blog](https://eng.uber.com/feed/)
- * [Spotify Blog](https://engineering.atspotify.com/feed/)
- * [Facebook Engineering](https://engineering.fb.com/feed/)
- * [LinkedIn Engineering](https://engineering.linkedin.com/blog.rss.html)
- * [evernote.com | Blog](https://evernote.com/blog/feed/)
- * [GitHub Blog](https://github.blog/feed/)
- * [Instagram Blog](https://instagram-engineering.com/feed)
- * [Instagram Engineering](https://instagram-engineering.com/feed)
- * [Pinterest Engineering](https://medium.com/feed/@Pinterest_Engineering)
- * [Airbnb Engineering](https://medium.com/feed/airbnb-engineering)
- * [DAZN Engineering](https://medium.com/feed/dazn-tech)
- * [Technology at PayPal](https://medium.com/feed/paypal-tech)
- * [Medium Enginnering](https://medium.engineering/feed)
- * [Netflix Blog](https://netflixtechblog.com/feed)
- * [Netflix TechBlog](https://netflixtechblog.com/feed)
- * [Slack Engineering](https://slack.engineering/feed/)
- * [Stack Overflow Blog](https://stackoverflow.blog/feed/)
- * [eBay Tech Blog](https://tech.ebayinc.com/rss/)
- * [Fastly Blog](https://www.fastly.com/g-assets/blog_rss.xml)
- * [Gitpod.io](https://www.gitpod.io/rss.xml)
- * [Pulumi Blog](https://www.pulumi.com/blog/rss.xml)
- * [Skypack Blog](https://www.skypack.dev/blog/feed/)
- * [Yahoo Engineering](https://yahooeng.tumblr.com/rss)
+- [Chromium Blog](http://blog.chromium.org/atom.xml)
+- [Google Developers Blog](http://feeds.feedburner.com/GDBcode)
+- [AWS News Blog](https://aws.amazon.com/jp/blogs/aws/feed/)
+- [Amazon Web Services Blog](https://aws.amazon.com/jp/blogs/news/feed/)
+- [Cloudflare Blog](https://blog.cloudflare.com/rss/)
+- [Atlassian Developer Blog](https://blog.developer.atlassian.com/feed/)
+- [Google Chrome Blog](https://blog.google/products/chrome/rss)
+- [Heroku](https://blog.heroku.com/engineering/feed)
+- [Twitter Blog](https://blog.twitter.com/engineering/en_us/blog.rss)
+- [WHATWG Blog](https://blog.whatwg.org/feed)
+- [Google Cloud Platform Blog Japan](https://cloudblog.withgoogle.com/ja/products/gcp/rss/)
+- [Google Developers Japan](https://developers-jp.googleblog.com/atom.xml)
+- [Uber Engineering Blog](https://eng.uber.com/feed/)
+- [Spotify Blog](https://engineering.atspotify.com/feed/)
+- [Facebook Engineering](https://engineering.fb.com/feed/)
+- [LinkedIn Engineering](https://engineering.linkedin.com/blog.rss.html)
+- [evernote.com | Blog](https://evernote.com/blog/feed/)
+- [GitHub Blog](https://github.blog/feed/)
+- [Instagram Blog](https://instagram-engineering.com/feed)
+- [Instagram Engineering](https://instagram-engineering.com/feed)
+- [Pinterest Engineering](https://medium.com/feed/@Pinterest_Engineering)
+- [Airbnb Engineering](https://medium.com/feed/airbnb-engineering)
+- [DAZN Engineering](https://medium.com/feed/dazn-tech)
+- [Technology at PayPal](https://medium.com/feed/paypal-tech)
+- [Medium Enginnering](https://medium.engineering/feed)
+- [Netflix Blog](https://netflixtechblog.com/feed)
+- [Netflix TechBlog](https://netflixtechblog.com/feed)
+- [Slack Engineering](https://slack.engineering/feed/)
+- [Stack Overflow Blog](https://stackoverflow.blog/feed/)
+- [eBay Tech Blog](https://tech.ebayinc.com/rss/)
+- [Fastly Blog](https://www.fastly.com/g-assets/blog_rss.xml)
+- [Gitpod.io](https://www.gitpod.io/rss.xml)
+- [Pulumi Blog](https://www.pulumi.com/blog/rss.xml)
+- [Skypack Blog](https://www.skypack.dev/blog/feed/)
+- [Yahoo Engineering](https://yahooeng.tumblr.com/rss)
 
 ### Article
 
-* [DevelopersIO](https://dev.classmethod.jp/feed/)
-* [InfoQ](https://feed.infoq.com/jp)
-* [MOONGIFT](https://feeds.feedburner.com/moongift)
-* [GIGAZINE](https://gigazine.net/news/rss_2.0/)
-* [INTERNET Watch](https://internet.watch.impress.co.jp/data/rss/1.0/iw/feed.rdf)
-* [Publickey](https://www.publickey1.jp/atom.xml)
-* [Thoughtworks insights](https://www.thoughtworks.com/rss/insights.xml)
+- [DevelopersIO](https://dev.classmethod.jp/feed/)
+- [InfoQ](https://feed.infoq.com/jp)
+- [MOONGIFT](https://feeds.feedburner.com/moongift)
+- [GIGAZINE](https://gigazine.net/news/rss_2.0/)
+- [INTERNET Watch](https://internet.watch.impress.co.jp/data/rss/1.0/iw/feed.rdf)
+- [Publickey](https://www.publickey1.jp/atom.xml)
+- [Thoughtworks insights](https://www.thoughtworks.com/rss/insights.xml)
 
 ### News
 
-* [WICG - Latest topics](https://discourse.wicg.io/latest.rss)
-* [WICG - Latest posts](https://discourse.wicg.io/posts.rss)
-* [InfoQ - News](https://feed.infoq.com/news/)
-* [IETF Tools releases](https://tools.ietf.org/tools/atomfeed.xml)
-* [O'Reilly Japan - New & Upcomming](https://www.oreilly.co.jp/catalog/soon.xml)
-* [W3C News](https://www.w3.org/blog/news/feed)
+- [WICG - Latest topics](https://discourse.wicg.io/latest.rss)
+- [WICG - Latest posts](https://discourse.wicg.io/posts.rss)
+- [InfoQ - News](https://feed.infoq.com/news/)
+- [IETF Tools releases](https://tools.ietf.org/tools/atomfeed.xml)
+- [O'Reilly Japan - New & Upcomming](https://www.oreilly.co.jp/catalog/soon.xml)
+- [W3C News](https://www.w3.org/blog/news/feed)
