--- packages/ui/CHANGELOG.md
@@ -1,19 +1,0 @@
-# @sb/ui
-
-## 0.0.4
-
-### Patch Changes
-
-- 4bd9dd2: Add SSG
-
-## 0.0.3
-
-### Patch Changes
-
-- Integration etc
-
-## 0.0.2
-
-### Patch Changes
-
-- 2b746fd: Sample
