--- packages/ui/package.json
@@ -1,73 +1,0 @@
-{
-  "name": "@sb/ui",
-  "version": "0.0.4",
-  "description": "Create a reusable Qwik component library",
-  "main": "./lib/index.qwik.mjs",
-  "qwik": "./lib/index.qwik.mjs",
-  "types": "./lib-types/index.d.ts",
-  "exports": {
-    ".": {
-      "import": "./lib/index.qwik.mjs",
-      "require": "./lib/index.qwik.cjs",
-      "types": "./lib-types/index.d.ts"
-    }
-  },
-  "files": [
-    "lib",
-    "lib-types"
-  ],
-  "packageManager": "pnpm@8.6.2",
-  "engines": {
-    "npm": "use pnpm please!",
-    "yarn": "use pnpm please!",
-    "pnpm": ">=8.6.2"
-  },
-  "private": false,
-  "type": "module",
-  "scripts": {
-    "build": "qwik build",
-    "build-storybook": "storybook build",
-    "build.lib": "vite build --mode lib",
-    "build.types": "tsc --emitDeclarationOnly",
-    "dev": "vite --mode ssr",
-    "dev.debug": "node --inspect-brk ./node_modules/vite/bin/vite.js --mode ssr --force",
-    "fmt": "prettier --write .",
-    "fmt.check": "prettier --check .",
-    "lint": "eslint \"src/**/*.ts*\"",
-    "release": "np",
-    "start": "vite --open --mode ssr",
-    "storybook": "storybook dev -p 6006",
-    "test": "vitest run",
-    "test.unit": "vitest components",
-    "test.unit.ui": "vitest --ui components",
-    "qwik": "qwik"
-  },
-  "devDependencies": {
-    "@builder.io/qwik": "1.2.6",
-    "@storybook/addon-essentials": "7.0.23",
-    "@storybook/addon-links": "7.0.23",
-    "@storybook/blocks": "7.0.23",
-    "@storybook/builder-vite": "7.0.23",
-    "@storybook/html": "7.0.23",
-    "@storybook/html-vite": "7.0.23",
-    "@types/eslint": "8.40.2",
-    "@types/node": "^20.4.0",
-    "@typescript-eslint/eslint-plugin": "5.61.0",
-    "@typescript-eslint/parser": "5.61.0",
-    "@vitest/ui": "^0.32.2",
-    "eslint": "8.44.0",
-    "eslint-plugin-qwik": "latest",
-    "np": "7.6.1",
-    "prettier": "2.8.8",
-    "storybook": "7.0.23",
-    "storybook-framework-qwik": "0.2.2",
-    "typescript": "5.1.6",
-    "undici": "5.22.1",
-    "vite": "4.4.8",
-    "vite-tsconfig-paths": "4.2.0",
-    "vitest": "^0.32.2"
-  },
-  "dependencies": {
-    "@qwikest/icons": "^0.0.8"
-  }
-}
