--- packages/ui/src/index.ts
@@ -1,3 +1,0 @@
-export { Logo } from "./components/logo/logo";
-export { Counter } from "./components/counter/counter";
-export { PostSummaryList } from "./components/post-summary-list/post-summary-list";
