--- packages/ui/src/root.tsx
@@ -1,17 +1,0 @@
-import { Counter } from "./components/counter/counter";
-import { Logo } from "./components/logo/logo";
-
-export default () => {
-  return (
-    <>
-      <head>
-        <meta charSet="utf-8" />
-        <title>Qwik Blank App</title>
-      </head>
-      <body>
-        <Logo />
-        <Counter />
-      </body>
-    </>
-  );
-};
